pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "optichain"
version = "0.1.0"
description = "OptiChain — Supply Chain Intelligence Platform (Forecasting + Inventory Optimization + MLOps)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.2",
  "numpy>=1.26",
  "scikit-learn>=1.5",
  "mlflow>=2.14",
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "prometheus-client>=0.20",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "httpx>=0.27",
  "ruff>=0.5",
  "black>=24.4",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
